+++ omop_alchemy/cdm/handlers/__init__.py
@@ -1,0 +1,5 @@
+from .vocabs_and_mappers import make_concept_resolver
+
+__all__ = [
+    "make_concept_resolver",
+]