+++ omop_alchemy/cdm/handlers/vocabs_and_mappers/concept_normalisers.py
@@ -1,10 +1,7 @@
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from .typing import Normaliser
-
-def normalize_default(s: str) -> str:
-    return s.strip().lower()
+    from .vocab_handlers import Normaliser
 
 def compose_normalizers(*fns: "Normaliser") -> "Normaliser":
     def _inner(s: str) -> str:
@@ -13,17 +10,28 @@ def compose_normalizers(*fns: "Normaliser") -> "Normaliser":
         return s
     return _inner
 
-def strip_uicc(code):
+"""
+This module contains core normalisation functions for concept lookups.
+
+Custom normalisers can be passed to LookupSpec instances to perform normalisation 
+on input values before lookup, and if similar functions are needed in multiple places,
+they can be added here and composed as needed.
+"""
+
+def normalize_default(s: str) -> str:
+    return s.strip().lower()
+
+def strip_uicc(code: str) -> str:
     return code.lower().replace('ajcc', 'ajcc/uicc')
 
-def make_stage(val):
+def make_stage(val: str) -> str:
     val = val.lower()
     roman_lookup = [('-iii', '-3'), ('-iv', '-4'), ('-ii', '-2'), ('-i', '-1'), ('nos', '')]
     for replacement in roman_lookup:
         val = val.replace(*replacement)
     return val
 
-def site_to_NOS(icdo_topog):
+def site_to_NOS(icdo_topog: str) -> str:
     split_topog = icdo_topog.split('.')
     if '.' not in icdo_topog:
         return f'{icdo_topog}.9'